refactor(FE): compoent style 분리 #57
