+++ client/components/main/mainsection/index.tsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import { useRecoilValue } from 'recoil';
+import Link from 'next/link';
+import ArticleCard from '../articlecard';
+import newsfeedState from '../../../atom/newsfeedState';
+import { ArticlesSection, FakeButton, NewArticleSection, Placeholder, Wrapper } from './index.style';
+
+export default function MainSection() {
+  const newsfeedList = useRecoilValue(newsfeedState);
+
+  return (
+    <Wrapper>
+      <Link href="/write">
+        <NewArticleSection>
+          <Placeholder>무슨 생각 하세요?</Placeholder>
+          <hr />
+          <FakeButton type="button">글쓰기</FakeButton>
+        </NewArticleSection>
+      </Link>
+      <ArticlesSection>
+        <ArticleCard />
+        {newsfeedList.map((item) => (
+          // eslint-disable-next-line no-underscore-dangle
+          <ArticleCard author={item.author} key={item._id} description={item.description} title={item.title} />
+        ))}
+      </ArticlesSection>
+    </Wrapper>
+  );
+}
