+++ client/components/main/sidebar/index.tsx
@@ -1,10 +1,8 @@
 import React, { useState, useRef, RefObject } from 'react';
-import styled from '@emotion/styled';
 import Link from 'next/link';
-import { displayColumn } from '../../styles/mixin';
-import Menu from './SideBar/Menu';
-import Title from './SideBar/Title';
-import COLORS from '../../styles/color';
+import Menu from './menu';
+import Title from './title';
+import { Dropdown, Setting, SideMenuBox, Wrapper } from './index.style';
 
 const menuList = [
   { routeSrc: '/', imgSrc: '/home.svg', text: '홈' },
@@ -46,50 +44,3 @@ export default function SideBar() {
     </Wrapper>
   );
 }
-
-const Dropdown = styled.div`
-  width: fit-content;
-  /* position: fixed; */
-  margin-left: 30px;
-  justify-content: center;
-  align-items: right;
-  margin-bottom: 10px;
-  div {
-    position: relative;
-    width: 100%;
-    height: 100%;
-    white-space: nowrap;
-    background-color: ${COLORS.WHITE};
-    border-radius: 10px;
-    ${displayColumn}
-    border: 1px solid ${COLORS.PRIMARY_DARK};
-    li {
-      list-style: none;
-      font-size: 14px;
-      margin: 10px 10px;
-    }
-  }
-`;
-
-const Wrapper = styled.aside`
-  width: ${({ theme }) => theme.sidebar.maxWidth};
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  background-color: ${({ theme }) => theme.sidebar.backgroundColor};
-  @media only screen and (max-width: ${({ theme }) => theme.smallWindow}) {
-    width: ${({ theme }) => theme.sidebar.minWidth};
-  }
-`;
-
-const SideMenuBox = styled.div`
-  width: 100%;
-  flex: 1;
-`;
-
-const Setting = styled.div`
-  width: 100%;
-  height: 50px;
-  margin-bottom: 20px;
-`;
